Parse OCaml line directives for BISECT comments
OCaml source code has line directives (# 1 "foo.ml") that change the
position, and more importantly for our case, filenames that are being
lexed. We would like to keep track of that information for correctly
implementing comments. The overall instrumentation though, is _still_
the top level filename because that strucutre is passed to the PPX
rewriter.