Reduce modes to one fast mode
Cleans up issues described in #41. Briefly, all the modes were
thread-safe due to the nature of the OCaml runtime, so all have been
replaced by the fastest mode.

As a consequence:
- The BisectThread module has been eliminated.
- The runtime has been greatly simplified to a register function
  (init_with_array) and a dump function called by at_exit.
- The old -mode argument now does nothing, but is kept for
  compatibility. A test was added to ensure it can be used.
- The bisect_ppx.fast package is now a duplicate of plain bisect_ppx.
- The big comment in runtime.mli has been updated to reflect the many
  changes since it was written.
- Several tests are no longer relevant and have been eliminated.

The reference outputs in ppx-integration have not been updated. They
will be updated once we are able to run those tests.

Resolves #41.