Apply Bisect_ppx to itself
Created a parallel build process, which first builds a copy of
Bisect_ppx whose runtime is named Meta_bisect instead of Bisect (to
avoid a name clash). Then, it builds a second copy of Bisect_ppx, with
every source file instrumented by Meta_bisect. This includes the syntax
extension, the report generator, and the runtime itself. Tests are then
run with this second instrumented Bisect_ppx. At the end of testing,
all the coverage files (close to 100 at the moment) are collected and
used for generating reports.

The instrumented Bisect_ppx is built by running make dev. This is now
required for testing, instead of the previous make all.

The regular build process for Bisect_ppx that is to be installed for
users is unaffected. That target has been renamed from make all to make
build.

To see coverage in action, run make unit in tests/ for a summary. Then,
optionally, run make coverage for a detailed HTML report.

The running time of performance tests is largely unaffected, because
Meta_bisect does not instrument programs under test except for the
runtime in module Bisect, and that module is used only during program
startup and exit.

Changes made to support this:
- Created additional intermediate directories _build.meta and
  _build.instrumented for the build processes, plus _install.meta, the
  latter for passing with -I to the compilers (using just
  -I _build.meta makes too many files visible, and results in duplicate
  module warnings).
- Created a module Extension. Meta_bisect differs from Bisect in that
  it generates *.out.meta files instead of *.out files. This allows
  them to be easily distinguished during testing, especially since both
  Bisect and Meta_bisect are linked into programs under test. Extension
  just contains the extension string, and is generated by Ocamlbuild.
- Added an option -runtime to the rewriter, so we can tell it to emit
  calls to Meta_bisect when instrumenting Bisect.

Resolves #36.