Install meta and instrumented Bisect as packages
These are now being installed in a little findlib directory _findlib/,
which replaces the previous _install.meta/. The main purpose of this
change is to make it possible to refer to both third-party ppx
rewriters under test and to the instrumented bisect_ppx using -package
arguments. This allows control over the order in which rewriters are
applied, which is necessary for ppx-integration tests. -dsource is not
sufficient because of bugs in -dsource and because it loses location
information, and it is better not to test with -dsource anyway since
that's not what users will use.

This change complicates the Makefile's install target, but simplifies
the testing code slightly.