Exhaustive matching in case instrumentation
Added an extra _ case to each match expression generated by Bisect_ppx
for marking points in cases of the user's match and function
expressions.

This ensures that if there is a bug in Bisect_ppx that makes the
generated match expression fail, user code will not receive an
unexpected exception.

This makes it more difficult to detect bugs in Bisect_ppx. For this
reason, generation of the extra _ case can be disabled by passing
-inexhaustive-matching to the instrumenter. This is done during the
report, thread safety, OUnit integration, and performance tests.

The other tests examine and do not run the instrumented code, so the
option would offer no benefit and would be counterproductive.