Flag for disabling points on or-pattern clauses
Passing -simple-cases to the instrumenter now prevents the generation
of separate points on or-pattern clauses, i.e. no secondary match
expressions are inserted into the code being instrumented. Behavior of
instrumented code is then almost the same as before the or-pattern
commit, the only difference being that the location of the point for
each case is at the beginning of the case's pattern, not at the
beginning of its expression.