Ocamlbuild plugins and usage test
Added a small library for use in Ocamlbuild plugins. This adds a new
tag "coverage". If the environment variable BISECT_COVERAGE is set to
"YES" while building, files with tag "coverage" get the option
"-package bisect_ppx". The accompanying usage test is meant to test
correct installation of, and linking with, Bisect_ppx itself and this
plugin library. The test's Makefile and myocamlbuild.ml show minimal
usage of the library.

The plugin library is packaged as a findlib package bisect_ppx.plugin.