Travis: test on 4.03 and improve matrix, script
Before this commit, we were testing on 4.02 on Ubuntu and whatever
version Homebrew had on OS X, which happened to be 4.03 - despite the
build matrix claiming 4.02.

Now explicitly testing on both 4.02 and 4.03 on both systems. The build
script now checks which version of the compiler is actually installed.

Removed the workaround for https://github.com/ocaml/opam/issues/1747.
That was fixed in https://github.com/ocaml/opam/pull/1756.

Remaining changes have to do with brevity (set -e instead of explicit
checks, set -x instead of echo) and the new build matrix (no need to
check for Linux when checking $COVERAGE).