fix instrumentation of refutation clauses `p -> .`
refutation clauses `| p -> .` must not get instrumented, as
instrumentation would generate code of the form

     | p -> <instrumentation>; .

that makes the type-checker fail with an error as it does not
recognize the refutation clause anymore:

    File "test.ml", line 3, characters 2-4:
    Warning 11: this match case is unused.
    File "test.ml", line 3, characters 8-9:
    Error: Unreachable expression was reached

The type-checker could in principle be extended to recognize
right-hand-sides of the form (e; .) as statically unreachable, but
this fails at least with OCaml 4.03 and 4.04.

Note that this patch could be trivially extended to automatically
consider clauses of the form

    | p -> assert false

as unreachable, which I would personally find desirable as they are
(in my code) always a clear intent of intentional unreachability
(so I would never write a test to cover them, and would have to
manually add an excluding comment). The current instrumenter does not
exclude these, so I did not implement it.