do not add marks in clauses of the form `| p -> assert false`
Those clauses are now considered unreachable and ignored without
asking the user to explicitly use BISECT-IGNORE.

Note that arbitrary `assert false` expressions are not automatically
ignored: if you write `(foo; assert false)`, then either `foo` is dead
code or you consider that this `assert false` is reachable, so it is
good that you are warned either way. But the clause form is a common
idiom to indicate static unreachability, so it makes sense to handle
it specifically.