Backwards-compatible packaging
This eliminates the new ppx_bisect package, and restores bisect_ppx as
the PPX, and bisect_ppx.runtime as the runtime library. The packaging is
compatible with both traditional builds and jbuilder.

The Ocamlbuild plugin is not yet included.

The .fast and .plugin deprecated packages are also not yet included.