Backwards-compatible Ocamlbuild plugin
This commit goes against best practices, and makes Ocamlbuild a depopt
of Bisect_ppx. In a future release, we will provide a dedicated
bisect_ppx-ocamlbuild package, and warn users to switch over to it. We
will remove the depopt in a subsequent release.

Technically, the depopt is a breaking change, since previously
installing Bisect_ppx transitively pulled in Ocamlbuild, so we always
provided the plugin. However, anyone using the plugin must already be
depending on, and explicitly using, Ocamlbuild, so hopefully there are
few users that only depend on it transitively (implicitly).