Delete test/README.md
Much of this file is out of date with the new Jbuilder build system. In
addition, the main feature of the test suite, which is preserving the
output in case of a failed diff, no longer works, and will have to be
reimplemented entirely anew over Jbuilder.

Given that, a completely fresh README will have to be written at a
future date.

[skip ci]