Packaging: make Ocamlbuild a regular dependency
Ocamlbuild is used for the Bisect_ppx Ocamlbuild plugin, which is being
factored out into a separate package in a future release. However, this
has not been done yet.

Making ocamlbuild a normal dependency in the meantime ensures both no
breakage and no recompilation for Bisect_ppx. The dependency will be
removed when factoring out the plugin is completed.