Inline InstrumentState into InstrumentPpx
InstrumentState was originally a separate module, containing common
logic needed by both the Camlp4 and the PPX modes of Bisect. Camlp4
support was removed in 359859869365234572b384bee5de9cf88934bf36.

It now makes more sense to put all the logic directly into
InstrumentPpx, so that a reader doesn't have to look in two files to
understand the instrumenter.

For easy review, this commit simply pastes the signature and definition
of InstrumentState into instrumentPpx.ml. It will be untangled, and
whitespace will be adjusted, in subsequent commits.