Docs: do not mention Ocamlbuild plugin package
This reverts the effect of a previous commit on doc/advanced.md:

  2008454de776246bac85d3467245259695658e6e

Specifically, we no longer recommend installing the package
bisect_ppx-ocamlbuild, because that package has not yet been released to
OPAM. We should restore that text when the release is happening.

Prompted by #150.

The package was factored out in #146.

[skip ci]