Support conditional instrumentation
This is a workaround for Jbuilder. Jbuilder lacks a good way to
conditionally disable PPXs.

When the option -conditional is passed to Bisect_ppx during
preprocessing, it checks for the BISECT_ENABLE environment variable. If
it is present and set to YES (case-insensitive), Bisect_ppx instruments
code as normal. If it is absent, or not set to YES, Bisect_ppx does not
modify the AST.

When -conditional is not passed, Bisect_ppx instruments as normal. This
behavior is backwards-compatible.

For typical usage, see the new usage test in

  test/usage/jbuilder-conditional

However, the -simple-cases option is just an example, and should not be
used for any production code.