Initialize the runtime on demand
When the Bisect_ppx runtime is linked in with -linkall, but no files are
instrumented, we would like to avoid generation of .out files, or
running any other code in the runtime.

To achieve this, the runtime only performs its side effects if at least
one instrumented file is initialized, i.e. calls
Bisect.Runtime.init_with_array.

This is part of the workarounds for Jbuilder.

The way runtime initialization has been made demand-based in this commit
is not necessarily the most beautiful or legible; it is meant to be
refactored during a future review of Bisect_ppx. For now, easy review is
the priority.