Simplify file, structure exclusion checks a bit
This is towards inlining the InstrumentState module (formerly a separate
file), and making the state clear and local to each mechanism that uses
it. In particular, this removes an unnecessary mutable list and replaces
it with a boolean flag.

The check for whether structures were inside an attribute or extension
point seems to have become redundant at some point. See

  https://github.com/aantron/bisect_ppx/pull/8/commits/94302532331ee0c67b39c74a8f00d9f388836615#r134090850