Docs: make finding generated files more likely
If bisect-ppx-report is not given an -I option, it looks for .ml files
for the HTML report in the current directory. This typically means it
can only find checked-in source files.

In case some source files are generated, they usually end up in a
_build/ directory (most build systems) or _build/default/ (for a typical
Jbuilder project). So, suggest adding those options in the instructions.

Resolves #158.

[skip ci]