Get rid of the "points for file" hash table
Bisect_ppx only ever puts one key-value pair in it: the points for the
current file it is processing. So, replace it with a simple list.

In a future commit, we should probably make it so that the state is
reset for each new AST mapper class instance that is created (i.e., that
a new state is created with it).