Delete attempt at point deduplication
In most cases, if the AST traversal wants to generate a point where one
already exists, because of two expressions sharing the same location, it
doesn't do so. However, there is no reason to favor the first expression
encountered over the second. It's an arbitrary order dependence.

After this change, both expressions are instrumented. If either is
executed, the same visible point is marked as visited, so we have to,
and have always had to, make an effort to assign different locations to
each expression.

This should be possible for the vanilla OCaml AST, but might be tricky
for ASTs emitted by third-party PPXs.

In future work, we may want to warn the user, or emit verbose debugging
output, for duplicate instrumented locations.