Remove -simple-cases option
There are no users of it in GitHub, and probably none offline. It was
originally added in #56, when working on enhanced instrumentation of
or-patterns. It was added as an escape hatch, however it is now clear
that it's not necessary to have.

The usage tests were using -simple-cases to test passing of an option to
the PPX. They have been changed to use -conditional for that purpose
instead.