Add a JSON report
This outputs something like:

```
{
  "file.ml": {
    "points": [{
      "offset": 0,
      "identifier: 5
    }, {
      "offset": 10,
      "identifier": 0
    }]
  },
  "file2.ml": {
    ..
  },
  ...
}
```

This is basically the same as the `dump` format, but
much easier to parse.

The per-file report is an object containing the `"points"`
key rather than just the points themselves, since this
will allow us to add other keys in the future without
breaking backwards compatibility (for example, to add per-line
output).

Downsides:

 - This brings in a dependency on atdgen and yojson
 - Atdgen's generated .ml files have compile warnings,
   so we can't build with `--dev`. I'm not sure how to
   work around this yet.