Verify that attributes are not instrumented
Adds an explicit test that attributes are not instrumented in .ml files.
We already have such a test for .mli files.

We should also test that extension points are not instrumented, but the
current testing setup simply calls the compiler on test sources, and
that results in uninterpreted extension errors. We should address that
during the next cleanup of the testing setup.