Fix check for existing bisectWXYZ.out files
Before this fix, Bisect would keep trying to create these files, with
different suffixes, if creating the previous ones failed for *any*
reason. This included the parent directory not existing, in which case
none of the files could be created anyway. This would send Bisect into
an infinite loop.

Now, Bisect continues trying to create files only if the previous files
could not be created specifically because they already existed.

Unfortunately, the Bisect runtime picks up a Unix dependency, because
the functions in Sys and Pervasives don't seem to provide a reliable way
of determining why file creation failed. The Sys_error exception raised
by open_out_gen carries a string as payload, which doesn't seem to be as
future-proof as the Unix.EEXIST payload of Unix_error from
Unix.openfile.

Resolves #175.