Skip Runtime.dump under JSOO based on the backend, rather than exceptions
Previously, Runtime.dump would be skipped based when running a
js_of_ocaml-compiled binary by catching a Failure exception that's
thrown when attempting to use runtime.unix_open to write the coverage
data.  However, the exact type of the exception that's thrown can be
either Failure, if the binary was compiled without an implementation
of the Js.Error type, or Js.Error otherwise.  Since the Failure
exception is not a reliable way of verifying that the binary was
compiled with js_of_ocaml, we check the backend explicitly using
Sys.backend_type.