Fix collation of subdirectories in HTML report
We have a project with top-level source files as well as subdirectories, and the
listing was being cut off before the first subdirectory. This came down to the
`partition_files` function, which did not correctly handle directories
containing both files and subdirectories. (It also returned the directories in
reverse order, which this patch also fixes.)